6.2 1416

+++ Assets/Scripts/局内/UI/角色UI/ShieldSlider.cs
@@ -24,7 +24,7 @@ public class ShieldSlider : MonoBehaviour
             GetComponent<UnityEngine.UI.Image>().fillAmount = 0;
             return;
         }
-        float shieldPercent = targetCharacter.currentShield / targetCharacter.maxHP;
+        float shieldPercent = (float)targetCharacter.currentShield / targetCharacter.maxHP;
         GetComponent<UnityEngine.UI.Image>().fillAmount = Mathf.Clamp01(shieldPercent);
     }
 }
